app/requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
Jinja2==3.1.2
keras==2.13.1
mysqlclient==2.2.0
numpy==1.24.3
Pillow==10.0.0
tensorboard==2.13.0
tensorflow==2.13.0
urllib3==1.26.16
